refactor(pdf): split four files back under the size limit
Moving the harness into src/bin/ made it production code for the file-size
gate, and the print dialog pushed windows.rs over. Four splits rather than
four baseline entries:

  bin/pdf_smoke.rs   -> main.rs + fixtures.rs + verify.rs
  renderer/windows   -> windows.rs + windows_print.rs
  renderer/mod.rs    -> mod.rs + sink.rs

Each seam is a real one. verify.rs is the part that decides pass/fail by
reading the produced PDF rather than trusting a return value. sink.rs holds
the only invariants worth the name — settle-once and settle-on-drop. The
print dialog renders into a VISIBLE window and hands control to the user,
where export renders off-screen and completes on its own.

useExportOperations.ts shrank 457 -> 446 when the two platform guards came
out; the baseline is lowered to lock that in.

One bug worth recording: inserting  immediately before
 put it BETWEEN the #[cfg(target_os = "macos")] attribute and the
module it was meant to gate, so sink became macOS-only and macos became
unconditional. macOS stayed green; the Windows cross-check caught it with 24
unresolved-objc2 errors.

Refs #1284
