+++ src-tauri/src/pdf_export/renderer/mod.rs
@@ -22,8 +22,6 @@
 //! @coordinates-with commands.rs — the only caller
 //! @module pdf_export/renderer
 
-use std::path::PathBuf;
-use std::sync::{Arc, Mutex};
 use std::time::Duration;
 use tauri::{AppHandle, Emitter};
 
@@ -31,6 +29,9 @@ use super::page_spec::PageSpec;
 use crate::command_error::{CommandError, ErrorCode};
 use crate::localized_error;
 use tokio::sync::oneshot;
+
+mod sink;
+use sink::RenderSink;
 
 #[cfg(target_os = "macos")]
 mod macos;
@@ -41,6 +42,8 @@ use macos_ops as platform;
 
 #[cfg(target_os = "windows")]
 mod windows;
+#[cfg(target_os = "windows")]
+mod windows_print;
 #[cfg(target_os = "windows")]
 use windows as platform;
 
@@ -53,57 +56,6 @@ use linux as platform;
 /// up to ~60s of run-loop ticks; double that to give the dispatcher headroom
 /// without leaving the user staring at a frozen export forever.
 const PDF_OPERATION_TIMEOUT: Duration = Duration::from_secs(180);
-
-/// The outcome channel a platform body must settle exactly once.
-///
-/// Passed IN rather than returned, because two of the three platforms cannot
-/// produce a result before the UI closure must return (ADR-PDF6): WebView2
-/// posts its completion to the creating thread's message pump and WebKitGTK
-/// emits a signal on the owning GLib context. Blocking the closure to wait for
-/// either one deadlocks the loop that would deliver it.
-///
-/// The sink also owns the temp HTML file, because it must outlive NAVIGATION,
-/// not the dispatch that started it (ADR-PDF4).
-pub(super) struct RenderSink {
-    tx: Mutex<Option<oneshot::Sender<Result<(), CommandError>>>>,
-    temp_html: PathBuf,
-}
-
-impl RenderSink {
-    fn new(tx: oneshot::Sender<Result<(), CommandError>>, temp_html: PathBuf) -> Arc<Self> {
-        Arc::new(Self {
-            tx: Mutex::new(Some(tx)),
-            temp_html,
-        })
-    }
-
-    /// Deliver the outcome and drop the temp file. A second call is ignored,
-    /// so a platform that both returns an error and fires its callback cannot
-    /// panic or double-report.
-    pub(super) fn settle(&self, result: Result<(), CommandError>) {
-        let taken = self.tx.lock().unwrap_or_else(|p| p.into_inner()).take();
-        if let Some(tx) = taken {
-            let _ = std::fs::remove_file(&self.temp_html);
-            let _ = tx.send(result);
-        }
-    }
-}
-
-/// If a platform body is dropped without settling — an unwind, or a callback
-/// that never fires and is released — report it immediately instead of making
-/// the caller wait out the full timeout for a result that can never arrive.
-impl Drop for RenderSink {
-    fn drop(&mut self) {
-        let taken = self.tx.lock().unwrap_or_else(|p| p.into_inner()).take();
-        if let Some(tx) = taken {
-            let _ = std::fs::remove_file(&self.temp_html);
-            let _ = tx.send(Err(localized_error!(
-                ErrorCode::Internal,
-                "errors.pdf.abandoned"
-            )));
-        }
-    }
-}
 
 /// Progress event payload.
 ///
@@ -267,6 +219,9 @@ pub async fn print_document(app: AppHandle, html: String) -> Result<(), CommandE
         // its panel is modal, while Windows and Linux settle once the dialog
         // has been SHOWN. None of the three can detect what the user then does
         // with it — that has always been the documented contract here.
+        #[cfg(target_os = "windows")]
+        windows_print::print_on_main_thread(&app_clone, &temp_html_str, &temp_dir_str, sink_clone);
+        #[cfg(not(target_os = "windows"))]
         platform::print_on_main_thread(&app_clone, &temp_html_str, &temp_dir_str, sink_clone);
         let _ = std::fs::remove_file(&temp_html_str);
     })
@@ -291,7 +246,3 @@ pub async fn print_document(app: AppHandle, html: String) -> Result<(), CommandE
         )),
     }
 }
-
-#[cfg(test)]
-#[path = "mod.test.rs"]
-mod tests;
